fix(core): Normalize model IDs for token counting
Use the canonical provider/model identifier at the runtime boundary so prebuilt models select the correct local tokenizer.
