+++ packages/@n8n/agents/src/runtime/__tests__/model-token-counter.test.ts
@@ -1,13 +1,14 @@
 import { describe, expect, it } from 'vitest';
 
-import { createModelTokenCounter } from '../../types/sdk/observation-log';
+import { createModelTokenCounter } from '../model/model-token-counter';
 
 describe('createModelTokenCounter', () => {
-	it('uses o200k for OpenAI and cl100k for other providers', async () => {
+	it('selects the encoding for the model family', async () => {
 		const text =
 			'こんにちは世界。ユーザーは毎週金曜日にレポートを送信します。مرحبا بالعالم — يجب إرسال التقرير يوم الجمعة. 🚀🔐🧪';
 
 		expect(await createModelTokenCounter('openai/gpt-5')(text)).toBe(37);
+		expect(await createModelTokenCounter('openai/gpt-4-turbo')(text)).toBe(68);
 		expect(await createModelTokenCounter('anthropic/claude-haiku-4-5')(text)).toBe(68);
 	});
