+++ packages/@n8n/agents/src/runtime/memory/memory-orchestrator.ts
@@ -25,14 +25,10 @@ import type {
 import { AgentEvent } from '../../types/runtime/event';
 import type { AgentPersistenceOptions, ExecutionOptions, RunOptions } from '../../types/sdk/agent';
 import type { AgentDbMessage } from '../../types/sdk/message';
-import {
-	estimateObservationTokens,
-	type ObservationLogScope,
-	type ObservationLogTaskKind,
-	type TokenCounter,
-} from '../../types/sdk/observation-log';
+import type { ObservationLogScope, ObservationLogTaskKind } from '../../types/sdk/observation-log';
 import type { AgentRuntimeConfig } from '../loop/agent-runtime';
 import type { AgentMessageList } from '../model/message-list';
+import { estimateObservationTokens, type TokenCounter } from '../model/model-token-counter';
 import type { BackgroundTaskTracker } from '../state/background-task-tracker';
 import type { AgentEventBus } from '../state/event-bus';
 import {
